refactor/FE: Refactor components structure
