+++ Front-end/src/app/components/ProductsList.tsx
@@ -1,5 +1,5 @@
 import { getProds } from "../lib/data";
-import ProductCard from "../ui/product-card";
+import ProductCard from "./ui/product-card";
 
 export default async function ProductsList({
   searchParams,
