+++ Front-end/src/app/components/ui/product-form.tsx
@@ -9,9 +9,18 @@ import {
 } from "@heroicons/react/24/solid";
 import Link from "next/link";
 import { useState } from "react";
+import useCartStore from "../../store/cart";
 
-export default function ProductForm({ stock }: { stock: number }) {
+export default function ProductForm({
+  stock,
+  productId,
+}: {
+  stock: number;
+  productId: string;
+}) {
   const { user } = useAuthContext();
+
+  const { cart, addToCart } = useCartStore();
 
   const [quantity, setQuantity] = useState(1);
   const [stockWarning, setStockWarning] = useState(false);
@@ -39,6 +48,10 @@ export default function ProductForm({ stock }: { stock: number }) {
     }, 500);
   };
 
+  const handleAddToCart = () => {
+    addToCart({ productId, quantity });
+  };
+
   return (
     <>
       {user?.email ? (
@@ -60,7 +73,10 @@ export default function ProductForm({ stock }: { stock: number }) {
                 <PlusIcon width={18} height={18} />
               </button>
             </div>
-            <button className="flex-[1_1_55%] rounded-lg bg-new-mint text-new-gray py-2 hover:bg-new-peach-90 duration-200">
+            <button
+              onClick={handleAddToCart}
+              className="flex-[1_1_55%] rounded-lg bg-new-mint text-new-gray py-2 hover:bg-new-peach-90 duration-200"
+            >
               Add to cart
             </button>
             <button
