+++ Front-end/src/app/components/ui/register-form.tsx
@@ -4,7 +4,7 @@ import { NextFont } from "next/dist/compiled/@next/font";
 import { useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { registerUser } from "../lib/actions";
+import { registerUser } from "../../lib/actions";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
 import { useAuthContext } from "@/contexts/AuthProvider";
