+++ Front-end/src/app/register/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from "../ui/register-form";
+import RegisterForm from "../components/ui/register-form";
 
 import { ptSerif } from "../layout";
 import Link from "next/link";
